[GStreamer] Move some DMABuf-related code from the player to a new DMABufUtilities file https://bugs.webkit.org/show_bug.cgi?id=276481
Reviewed by Xabier Rodriguez-Calvar.

Refactoring aiming to reduce a bit the size of MediaPlayerPrivateGStreamer::pushDMABufToCompositor()
and improve readability.

* Source/WebCore/platform/SourcesGStreamer.txt:
* Source/WebCore/platform/graphics/gstreamer/DMABufUtilities.cpp: Added.
(WebCore::DMABufDestination::DMABufDestination):
(WebCore::DMABufDestination::~DMABufDestination):
(WebCore::DMABufDestination::copyPlaneData):
(WebCore::fillSwapChainBuffer):
* Source/WebCore/platform/graphics/gstreamer/DMABufUtilities.h: Added.
(WebCore::dmaBufFourccValue):
(WebCore::dmaBufColorSpaceForColorimetry):
(WebCore::DMABufDestination::mappingLock):
* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::pushDMABufToCompositor):
(WebCore::fourccValue): Deleted.
(WebCore::colorSpaceForColorimetry): Deleted.

Canonical link: https://commits.webkit.org/280906@main
